Switch to implot + vulkan for rendering

+++ include/2fish/models/market_snapshot.h
@@ -4,8 +4,8 @@
 #include <cstdint>
 
 struct MarketSnapshot {
-	std::array<long double, 101> bids_{};
-	std::array<long double, 101> asks_{};
+	std::array<double, 101> bids_{};
+	std::array<double, 101> asks_{};
 
 	uint64_t last_message_{}; // milliseconds
 	uint64_t last_updated_{}; // milliseconds
